Implemented a variant of the Visitor pattern (Design Patterns, GoF, 1995, p. 331) for SBML objects by adding the canonical accept() method.
Added const versions of the getListOfXXX() methods.

Renamed the methods:

  - getReactantById() -> getReactant()
  - getProductById()  -> getProduct ()

The naming scheme takes advantage of C++ overloading (i.e. you can
specify either a string id or an integer position to get the nth
reactant or product) and is consistent with similar methods elsewhere
in libSBML.  The "ById" portion of the name was a holdover from the C
API and was introduced by mistake when converting the C API to C++.