Added a Compartment to the Model, so that Constraint 1002:
  A Model that has a Species must also have at least one Compartment
  (L2v1 Section 4.5)

would not fail.  In general, only one constraint may fail per test file
and if multiple constraints are allowed to fail (indicated by the second
set of digits in the filename), they must all have the same constraint id
(and that constraint id is indicated by the first set of digits in the
filename).