Fixed undefined symbols error in Python binding with non-system-installed Python on MacOS X (10.3 or later).
According to configure.in of Python source code, -undefined dynamic_lookup
should be used for 10.3 or later. Actually, the option is needed to avoid
undefined symbols error when building a Python binding library with non-system-installed
Python on 10.3 or later.
Also, according to the man page of ld, environment variable MACOSX_DEPLOYMENT_TARGET
must be set to 10.3 or higher to use -undefined dynamic_lookup.
Currently, the environment variable is set in src/bindings/python/Makefile.in.