The bug involving bad schemaLocation turned out to be caused by more than one issue.  One issue was that libsbml's Xerces parser code called getLine()/getColumn() when (I guess) it shouldn't be calling them because Xerces in that situation doesn't report a valid line number.  On some systems (e.g., sbml.org), the end result was a seg fault, whereas on others (my Macos laptop) the numbers returned are merely crazy.
We don't have a way to communicate to a caller that the line/column
numbers may be junk.  We can't use 0 because that gets used to
indicate something else by XMLError.

So here is the new procedure: If the line and column numbers returned
by libSBML are equal to the C/C++ constant ULONG_MAX, then assume that
the numbers are junk and do not print line/column numbers as part of
the error reported by the validator.

Why?  Our line/column number values are unsigned integers, which thus
only allows 2 states to be communicated. (Explanation: you could use 0
to communicate an invalid state and let all other values indicate a
valid state.  If we had signed integers, we could communicate 3 state
values, using negative, 0, and positive values for the different
states.)  Unfortunately we can't use 0 because that value is already
being used to communicate a different meaning between users of
XMLError objects.

The change is a bit of a hack, but not completely unrealistic: we use
the largest possible value of the data type used to represent
line/column numbers.  (It's not completely unrealistic, because the
probability that an error in an file will involve these specific
numbers is vanishingly small.  Thus, we can say that if user code
encounters these values as line/column numbers, they should assume
they are actually meaningless and no valid line/column number could be
provided by the parser.)

The change involves the addition of

#define LIBSBML_UNKNOWN_LINE    ULONG_MAX
#define LIBSBML_UNKNOWN_COLUMN  ULONG_MAX

to XMLError.h and changing src/xml/XercesParser.cpp to use these
numbers when the error is among a certain set of error codes returned
by Xerces.

The other parser interfaces still need to be similarly modified.