Fixed tracker issue #2146714: syncAnnotation function improved.
syncAnnotation() functions in SBase, Model, and SpeciesReference class
could internally delete an old annotation object (XMLNode*) in an SBase 
object.
The problem is that syncAnnotation() functions are internally invoked
when getAnnotation() function called and this could lead to an invalid 
pointer to an already deleted old annotation object.

To avoid such an invalid pointer, an annotation object is now never 
deleted but just updated by using an assignment operator.