Tracker issue #2219230
The setProgramName function on the SBMLWriter class was not producing the comment as documented.

This is now implemented.

Note: the code to write the comment on the XMLOutputStream uses the function to return the libsbml version
from the file: common/libsbml-version.h. This means that any test code that uses XMLOutputStream must link
to the files in the common directory - hence the additions to the Makefiles in test directories.