Changed the number lexer to return numbers in e-notation instead of just as doubles.
Also fixed a bug in the 'translateMath' example that would crash if you gave it something you couldn't parse.

Also also, the new L3 parser is now tested like the old L1 parser was.  The tests are all identical, save a few newly-added ones, and the fact that what used to come across as '-1' now comes across as minus(1).