Added a L3ParserSettings object that *doesn't* inherit from the converter settings, because the the tradeoff of a the benefit of a better user interface didn't seem worth the price of a more confusing programmatic interface, since I only anticipate programmers using this interface anyway.
That said:  there is now a way to parse L3 formulae with any of three settings for what 'log(x)' turns into, and either collapsing or expanding unary minuses, not only on numbers but on anything:  "--(whatever)" reduces to "(whatever)".  AND there are tests in TestL3FormulaParser that test this!  AND the old "SBML_parseL3Formula("string", model)" calls still work, since there is a L3ParserSettings constructor that takes only a Model*.  I am particularly proud of this.

If someone wants to go through and add inheritance from the conversion settings objects to this, I am OK with that as long as the existing API does not change (and is only added to).  But I do think it'll make for a more confusing object.