Reverted and fixed many of Sarah's changes:  the new functions had the old signature, but the new behavior, which would have caused even more obscure bugs than simply changing the signature would have.  Also, finally noticed at the very end that many of the functions were indeed private to begin with, meaning that the whole folderol wasn't needed after all, but hey.
The big change here is twofold:  one, all removal of deleted and replaced items happens after everything else, and takes great pains to not double-delete anything.  This allows the possibility of replacing the child of a deleted element (the single example that makes any sense in this regard is replacing a SpeciesReference child of a deleted Reaction).  Secondly, all the old file signatures are replaced and have their old behavior, and are marked as being deprecated (in that an error is introduced into the manipulated file saying 'hey, you're doing it wrong').

mToRemove has been removed again, as it was not needed in the old behavior, nor was it needed with the new behavior and the new function signatures.

Instead of perform*, the new functions are now named collect*, as they don't actually delete/replace anything.

Removed the mToRemove and mRemoved from CompBase, as neither served any function any more.

Moved the now-deprecated ReplacedBy::performReplacement and ReplacedElement::performReplacement to Replacing, as both were now the same.