so took out parsing of the math things removed from arrays actually the parser stayed the same but the package tells it that things are no valid (mainly because they might come back !)
I will add in the new things that are functions since these will just fall out of the code so long as it recognises the name. 

Thing like uplimit/lowlimit will be trickier but I suggest we hold off on these for now since no one is implementing more than vector and selector yet anyway.